Move frequency from UserTerm into service layer

+++ src/main/java/ubc/pavlab/rdp/settings/ApplicationSettings.java
@@ -151,7 +151,7 @@ public class ApplicationSettings {
     private OrganSettings organs;
 
     private boolean sendEmailOnRegistration;
-    private int goTermSizeLimit = 100;
+    private long goTermSizeLimit = 100L;
     public List<String> enabledTiers;
 
 }
